Split phase plan evidence audits
