+++ tests/docs_truth/phase_audit.rs
@@ -1,5 +1,9 @@
 #[path = "phase_audit/completion_audit.rs"]
 mod completion_audit;
+#[path = "phase_audit/generated_c_evidence.rs"]
+mod generated_c_evidence;
+#[path = "phase_audit/generic_diagnostics_evidence.rs"]
+mod generic_diagnostics_evidence;
 #[path = "phase_audit/imported_generic_diagnostics.rs"]
 mod imported_generic_diagnostics;
 #[path = "phase_audit/phase_plan.rs"]
